Rename the Legal Statement to Terms And Conditions. Updated the Cookies To Cookies Policy. Updated the LinkedIn link URL in Social Media links.
