+++ Constants/HomePage/Footer/SocialIconsList.ts
@@ -1,12 +1,12 @@
 import { SocialIconType } from "@/Interfaces/Home/Footer/SocialIconType";
 
 export const SocialIconsList: SocialIconType[] = [
-  { src: "/assets/Telegram.png", hoverSrc: "/assets/Telegram-h.png" },
-  { src: "/assets/Instagram.png", hoverSrc: "/assets/Instagram-h.png" },
-  { src: "/assets/Medium.png", hoverSrc: "/assets/Medium-h.png" },
-  { src: "/assets/Linkedin.png", hoverSrc: "/assets/Linkedin-h.png" },
-  { src: "/assets/Facebook.png", hoverSrc: "/assets/Facebook-h.png" },
-  { src: "/assets/Twitter.png", hoverSrc: "/assets/Twitter-h.png" },
-  { src: "/assets/YouTub.png", hoverSrc: "/assets/YouTub-h.png" },
-  { src: "/assets/Twitch.png", hoverSrc: "/assets/Twitch-h.png" },
+  { src: "/assets/Telegram.png", hoverSrc: "/assets/Telegram-h.png", link:"#" },
+  { src: "/assets/Instagram.png", hoverSrc: "/assets/Instagram-h.png", link: "#" },
+  { src: "/assets/Medium.png", hoverSrc: "/assets/Medium-h.png", link: "#" },
+  { src: "/assets/Linkedin.png", hoverSrc: "/assets/Linkedin-h.png", link:"https://www.linkedin.com/company/medmonk-inc-/" },
+  { src: "/assets/Facebook.png", hoverSrc: "/assets/Facebook-h.png", link: "#" },
+  { src: "/assets/Twitter.png", hoverSrc: "/assets/Twitter-h.png", link: "#" },
+  { src: "/assets/YouTub.png", hoverSrc: "/assets/YouTub-h.png", link: "#" },
+  { src: "/assets/Twitch.png", hoverSrc: "/assets/Twitch-h.png", link: "#" },
 ];
