+++ Interfaces/Home/Footer/SocialIconType.ts
@@ -1,4 +1,5 @@
 export interface SocialIconType {
   src: string;
   hoverSrc: string;
+  link: string;
 }
