+++ components/Home/Footer/Cookies.tsx
@@ -3,7 +3,7 @@ import { cookiesPolicy } from "@/Constants/Cookies/cookiesPolicy"
 const Cookies  = ()=>
     {
         return (
-            <section className="max-w-4xl mx-auto p-4 bg-white shadow-lg rounded-lg my-6 text-gray-700 my-28">
+            <section className="max-w-4xl mx-auto p-4 bg-white shadow-lg rounded-lg text-gray-700 my-28">
             <h1 className="text-2xl font-bold text-center mb-4">
               {cookiesPolicy.Title}
             </h1>
