+++ components/Home/Footer/TermsAndConditions.tsx
