+++ pages/TermsAndConditions/index.tsx
@@ -1,11 +1,10 @@
 import Layout from "../Layout";
 import { GetServerSideProps } from "next";
-import { MetaProps } from "@/Interfaces/SEO/MetaProps";
-import PageNotFound from "@/components/PageNotFound";
 import { getMetaTags } from "@/services/SeoService";
-import TermsAndConditions from "@/components/Home/Footer/LegalStatement";
+import { MetaProps } from "@/Interfaces/SEO/MetaProps";
+import TermsAndConditions from "@/components/Home/Footer/TermsAndConditions";
 
-const LegalStatement = ({ metaTags }: { metaTags: MetaProps }) => {
+const TermsAndConditionsPage = ({ metaTags }: { metaTags: MetaProps }) => {
   return (
     <>
       <Layout metaData={metaTags}>
@@ -18,4 +17,4 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
   const meta: MetaProps | null = await getMetaTags("LegalStatementPage");
   return { props: { metaTags: meta } };
 };
-export default LegalStatement;
+export default TermsAndConditionsPage;
